Move the beans panel to its own category
