+++ control-panel-management/src/main/resources/micronaut-control-panel/micronaut-control-panel.properties
@@ -1,15 +1,24 @@
-micronaut.control-panel.panels.beans.title = Bean Definitions
-micronaut.control-panel.panels.beans.icon = fa-plug
-micronaut.control-panel.panels.beans.order = 30
-
-micronaut.control-panel.panels.env.title = Environment Properties
-micronaut.control-panel.panels.env.icon = fa-sliders
-micronaut.control-panel.panels.env.order = 10
+# Main category
 
 micronaut.control-panel.panels.health.title = Application Health
 micronaut.control-panel.panels.health.icon = fa-laptop-medical
 micronaut.control-panel.panels.health.order = 0
 
+micronaut.control-panel.panels.env.title = Environment Properties
+micronaut.control-panel.panels.env.icon = fa-sliders
+micronaut.control-panel.panels.env.order = 10
+
 micronaut.control-panel.panels.loggers.title = Loggers
 micronaut.control-panel.panels.loggers.icon = fa-file-lines
 micronaut.control-panel.panels.loggers.order = 40
+
+
+# Beans category
+
+micronaut.control-panel.panels.beans.title = Bean Definitions
+micronaut.control-panel.panels.beans.icon = fa-plug
+micronaut.control-panel.panels.beans.order = 0
+
+micronaut.control-panel.panels.disabled-beans.title = Disabled Beans
+micronaut.control-panel.panels.disabled-beans.icon = fa-plug-circle-minus
+micronaut.control-panel.panels.disabled-beans.order = 10
